add an option to report more elaborate information if a run failed. Useful for debugging.

+++ sepp-package/run-sepp.sh
@@ -1,7 +1,7 @@
 #!/bin/bash  
 
 if [ $# -lt 2 ]; then
-   echo USAGE: $0 "[input fragments file] [output prefix] [optional: -x number-of-cores ] [optional: -A alignment subset size] [optional: -P placement subset size] [optional: any other SEPP argument] [optional: -t filename reference phylogeny] [optional: -a filename reference alignment]
+   echo USAGE: $0 "[input fragments file] [output prefix] [optional: -x number-of-cores ] [optional: -A alignment subset size] [optional: -P placement subset size] [optional: any other SEPP argument] [optional: -t filename reference phylogeny] [optional: -a filename reference alignment] [optional: -b 1 = report debugging information ]
    Optional commands need not be in order. Any SEPP option can also be passed. For example, use
    -x 8
    to make SEPP us 8 threads"
@@ -34,9 +34,23 @@ fi
 
 # from http://stackoverflow.com/a/2130323
 function cleanup {
-  echo "Removing $tmp"
-  rm -r $tmp
-  rm -r $tmpssd
+  exitcode=`echo $?`
+  if [ ! -z "$printDebug" ];
+  then
+    echo "========= Execution of SEPP failed with exit code $exitcode =================";
+    echo "temporary working directories are NOT deleted for further inspection:";
+    echo "  \$tmp = $tmp";
+    echo "  \$tmpssd = $tmpssd";
+    echo "--------- Content of STDOUT -----------------------------------------";
+    cat sepp-$name-out.log
+    echo "--------- Content of STDERR -----------------------------------------";
+    cat sepp-$name-err.log
+    echo "=====================================================================";
+  else
+    echo "Removing $tmp";
+    rm -r $tmp
+    rm -r $tmpssd
+  fi
   unset TMPDIR
 }
 trap cleanup EXIT
@@ -74,6 +88,10 @@ do
 			;;
 		-t|--referencePhylogeny)
 			t="$2"
+			shift # past argument
+			;;
+		-b|--debugInformation)
+			printDebug="$2"
 			shift # past argument
 			;;
 		*)
